+++ src/main/java/se/su/dsv/pvt/utemaning/backend/Location.java
@@ -5,19 +5,19 @@ package se.su.dsv.pvt.utemaning.backend;
  */
 public class Location {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
 
-    public Location(int x, int y){
+    public Location(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
-    public int getY() {
+    public double getY() {
         return y;
     }
